requirements.txt
asgiref==3.8.1
build==1.2.1
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
cryptography @ file:///home/builder/.termux-build/python-cryptography/src
distlib==0.3.8
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
idna==3.7
importlib_metadata==8.0.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
jwcrypto==1.5.6
keyring==24.3.1
more-itertools==10.3.0
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
rapidfuzz==3.9.4
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
sqlparse==0.5.0
tomlkit==0.12.5
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
zipp==3.19.2
